backend/requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ─── Computer Vision ──────────────────────────────────────────────────────────
opencv-python-headless==4.9.0.80
numpy==1.26.4

# ─── Face Analysis (InsightFace = RetinaFace detector + ArcFace embedder) ─────
insightface==0.7.3
onnxruntime==1.18.0          # CPU-only; replace with onnxruntime-gpu for CUDA
# onnxruntime-gpu==1.18.0    # Uncomment for GPU acceleration

# ─── Vector Search ────────────────────────────────────────────────────────────
faiss-cpu==1.8.0              # CPU FAISS; use faiss-gpu for CUDA systems

# ─── Utilities ────────────────────────────────────────────────────────────────
Pillow==10.3.0
aiofiles==23.2.1
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1
